requirements.txt
anyio==3.7.1
bcrypt==4.0.1
bidict==0.22.1
boto3==1.26.160
botocore==1.29.160
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==1.16.0
elastic-transport==8.4.0
elasticsearch==8.8.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
gevent==23.7.0
gevent-websocket==0.10.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
pluggy==1.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
PyMySQL==1.0.3
pytest==7.3.1
python-dateutil==2.8.2
python-decouple==3.8
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
yapf==0.33.0
zope.event==5.0
zope.interface==6.0
